+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/programfile/PackageInfoWriter.java
@@ -541,7 +541,7 @@ public class PackageInfoWriter {
             case LINE_NUMBER_TABLE_ATTRIBUTE:
                 LineNumberTableAttributeInfo lnNoTblAttrInfo = (LineNumberTableAttributeInfo) attributeInfo;
                 LineNumberInfo[] lineNumberInfoEntries = lnNoTblAttrInfo.getLineNumberInfoEntries();
-                attrDataOutStream.writeShort(lineNumberInfoEntries.length);
+                attrDataOutStream.writeInt(lineNumberInfoEntries.length);
                 for (LineNumberInfo lineNumberInfo : lineNumberInfoEntries) {
                     writeLineNumberInfo(attrDataOutStream, lineNumberInfo);
                 }
